CCCP-388, fix: resolve comment
